[Nokia-IXR7250E] disable pcie aer (#18864)
In Linux kernel v6.0, commit, f26e58bf6f5470 "PCI/AER: Enable error reporting when AER is native" changed the default behavior of Advanced Error Reporting (AER). Reporting is now enabled by default, which was not the case prior. There are custom FPGA devices on our PCI bus that are not yet configured nor enabled across reboot until the Sonic platform is initialized. Therefore this is impacting AER adversely. Since AER was not enabled in 202205, this is not a regression.