[baseimage]: Added wireless-regdb as package to install for Debian (#18650)
In Debian 6.1.38, rfkill is used, which uses cfg80211. cfg80211 will attempt to load on boot and will look for regulatory.db, which is installed via wireless-regdb, a missing package.

It looks like the cause for the cfg80211 module getting automatically loaded is mentioned in systemd/systemd#27520 (specifically, systemd/systemd#20802). Some code refactoring occurred in systemd-networkd that causes the module to automatically get loaded, which then looks for the regulatory.db firmware file. (Also, the cfg80211 module depends on the rfkill module, not the other way around.)

systemd-networkd isn't enabled by default by Debian, but looks like SONiC now uses systemd-networkd in one case, starting with #18178, which is probably when this issue started happening. I do see the advantages of using it more; network configuration could potentially become a bit easier with this, but it would take some effort in migrating everything over to it.

Given all of this, either the cfg80211 module needs to be blacklisted, or the wireless-regdb package needs to be added. Considering there could be a theoretical use case of someone trying to use SONiC with wireless interfaces (although it would be clunky), it might be fine to add the wireless-regdb package.