[dhcp_server] Add support for monitoring used but invalid range/option (#18945)
Why I did it
Currently, dhcpservd wouldn't monitor option / range change which is referred in DHCP_SERVER_IPV4 table but is invalid or not been defined. It would miss config when they are defined or become valid.

How I did it
Treat invalid or not defined option / range as interested items.