[multi-asic] Use asic_id/asic_name for APIs to get port config data (#18452)
Fixes: https://github.com/sonic-net/sonic-buildimage/issues/18431

#### Why I did it
sonic-cfggen misses handling for namespace at couple of locations. Due to this, when port config is fetched, it returns data from global config and not the namespace.

The function calls for
get_path_to_port_config_file
get_port_config

needs to be fixed to handle namespace for multi-asic.

#### How I did it
1.Pass asic_id as argument when calling get_path_to_port_config_file()
For single asic, asic_id is None and the function returns the correct port_config.ini file path.
For multi-asic, using asic_id provides the correct file path for the asic.

2. Pass asic_name as argument when calling get_port_config()

#### How to verify it
sonic-cfggen -H --print -n -k
This command should return the correct port config for the namespace.