[Mellanox] Verify the value of the control type sysfs before changing it to FW control (#19046)
- Why I did it
On Mellanox platforms, changing the control type sysfs value is not supported for ports that are set to firmware control. Therefore, attempting to change this value to firmware control will result in an error log, which we want to avoid.

- How I did it
I added a check to ensure that the control sysfs value is set to firmware control only if it is not already in firmware control.

- How to verify it
Ensure that after a warm reboot on the Mellanox platform (an attempt to change the sysfs value from fw_control to fw_control is only applicable in this scenario), the following error does not appear:
sxd_kernel: [error] sx_core_set_module_control: Failed control change: module is dependent (FW control).