[database] Do not run local a Redis instance for the remote database (#19238)
The local Redis daemon should not run for a remote database.

How I did it
Do not generate a supervisord configuration entry for the remote database instances.

How to verify it
Compile and run the DPU image. The database container on the DPU should run Redis daemons only for local DB instances.