[docker-ptf] create symlinks for py3-only image (#19813)
Why I did it
The mix of docker-ptf with Python 2 + Python 3 (in virtual-env) and Python 3 images only across branches (with possibly some test scripts in sonic-mgmt not migrated to Python 3 in older branches) causes different path references to Python and PTF binaries. This can cause backporting a hassle. To prevent this the PR creates virtual environment (/root/env-python3) path to the real python 3 and PTF binaries. This keeps the paths same across docker-ptf images.

How I did it
Created symlinks for python and PTF from /root/env-python3/bin to /usr/bin.

How to verify it
Image built successfully and manually verified the existence of the links.