Remove default hash-offset for chassis-packet (#19745)
Why I did it
Fixes #19746

switch.json in swss docker sets the default hash offset to 0 for chassis-packet. The change was recently introduced via #18912 which was supposed to be only for T0/T1. On chassis-packet, each asics needs to set an unique hash offset to set an unique offset for load balancing algorithm. The new change overwrites platform setting and resets the hash-offset to 0 on all asics. This breaks the ecmp load balancing algorithm on chassis-packet.

How I did it
Remove the default hash-offset setting for chassis-packet. This setting is provided by platform based on platform specific rules.

How to verify it
Run sonic-mgmt fib/test_fib.py

Signed-off-by: anamehra <anamehra@cisco.com>