[Mellanox] Add dpuctl services to support "dark mode". (#19931)
- Why I did it
Smart Switch Dark Mode allows disabling of all DPUs, removing any influence the DPU has on the system.

- How I did it
Add the dpuctl.service that runs on system start and shuts down DPUs if needed.

- How to verify it
Compile the image and run it on the Nvidia Smart Switch. Verify that all DPUs are down after the system starts.