[Mellanox] Fix retry logic on discovery of MST device (#20389)
- Why I did it
Fixing retry logic when MST device is discovered. The current implementation only fetches the name of the device but doesn't verify if the device is accessible which can be confirmed by querying the device and ensuring the command passes.

- How I did it
Added a query command with device as parameter and ensured it passes.

- How to verify it
Running upgrade tests.