Platform DCS-7060DX5-64S misc update (#20323)
Why I did it
This is to add the HWSKU DCS-7060DX5-64S misc update. This is a must have for the platform support. This should get into the master branch as soon as possible and back port to the needed releases.

Work item tracking
Microsoft ADO (number only):
How I did it
Adding additional media, port config settings etc needed for the HWSKU.
This is an updated version of an older PR #19109 for 202305 which is closed. This PR should replace the older PR#19109 and should be cast to all active releases. This PR has the latest chip configuration base on Arista lab setting and the most up to date SI setting for optical transceivers.

How to verify it
We have verified it in house using the sonic mgmt test.
In additional, we have verify all the SI value which match our HW database for optical deployment.
This changes has been verified in Arista lab with 202305 , 202311, 202405 releases.