[bgp]: Tuning zebra nexthop-group keep parameter (#21135)
Why I did it
Setting the nexthop-group keep parameter to 1. This will instruct zebra not to save nexthop group for more than 1 second after removal. Without this zebra will keep nexthop group in the system for 180 seconds.
In scaled scenarios when this parameter is not set it resulted in the queue growing so big and crashing zebra due to OOM when there is test on link flapping.

How I did it
Update the zebra template and initialize nexthop-group keep as 1.

How to verify it
Running the scale test with link flapping and ensure no memory increase in zebra.