DellEMC : Z9432F determine reboot cause changes (#20984)
Why I did it
The reboot cause is not properly determined after each reboot. So the reboot history is also not maintained.

How I did it
The failure in determining the reboot cause is due to pcisysfs.py script failure in reading the registers.
The pcisysfs.py script used was using the old python2 format which were failing.
Modified the install scripts to use the latest pcisysfs.py for register read and write.