Banner: Remove dependency on sshd service (#21264)
- Why I did it
Fixes #19661

- How I did it
Removed the dependency on sshd service. Banner service will save the config data into the file after setting the config, so it will be available after the boot in any case

- How to verify it
Reboot the switch and wait for the serial prompt. SSH should become available after a couple of seconds.

Signed-off-by: Yevhen Fastiuk <yfastiuk@nvidia.com>