Dell:Initial commit for Z9664F platform (#20768)
Why I did it
To add support for Z9664F platform

How I did it
Implemented the support for the platform Z9664F

Switch Vendor: Dell
Switch SKU: Z9664F
ASIC Vendor: Broadcom
SONiC Image: sonic-broadcom.bin

How to verify it
Verified the platform show commands and also executed the sonic-mgmt testcases.
logs.txt

Added PDDF changes as well and attaching the logs
The syncd is not up and will be raising it to broadcom for the same as it requires SAI support.
logs.zip