Count the number of undeliverable IPinIP packets (#21811)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
When an IPinIP packet lands on a TOR where the inner destination IP is not learned, that packet gets dropped. Keep count of the number of dropped packets so we can monitor/detect any traffic issues.

##### Work item tracking
- Microsoft ADO **24106234**:

#### How I did it
Create a new table in COUNTERS_DB to store undeliverable IPinIP packets and update it whenever tunnel packet handler sniffs such a packet.

#### How to verify it
Send an IPinIP packet to a TOR where the inner destination IP of the packet is not resolved on the TOR. Verify that the `COUNTERS:IPINIP_TUNNEL` table roughly reflects the number of packets sent. We allow some inaccuracy in the counter due to poor scapy perfomance under load as being able to track trends is more important that a 100% accurate packet count.

Link to sonic-mgmt test: https://github.com/sonic-net/sonic-mgmt/pull/14128

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111
- [ ] 202205
- [ ] 202211
- [ ] 202305

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)