adding free radius (#21307)
Why I did it
In a subsequent PR in the sonic-mgmt repo, I will add RADIUS tests, and there needs to be a consumable RADIUS server in the infrastructure.

How I did it
Added the freeradius package the ptf docker j2 file.

How to verify it
I build the new docker image locally and then ran few succesfull app-topo tasks inside of sonic-mgmt. Also verifed that freeradius is installed.