Build fix for ptf-py3 issue on bullseye (#21886)
The latest version of setuptools-scm i.e. 8.2.0 does not support bdist_wheel. Hence there is a failure in target docker-syncd-rpc.gz due to ptf-py3 build failure. On further investigation found that bookworm installs 8.0.4 version of this package. Hence added this to the current Dockerfile.j2 of sonic-slave-bullseye. This will make it consistent with the earlier build process for ptf-py3 which used 8.1.0 version of setuptools-scm and also the sonic-slave-bookworm

Why I did it
To fix build failure of ptf-py3 on Bullseye

How I did it
Investigated the build failure and found that the recent update i.e. 8.2.0 of setuptools-scm python package does not support bdist_wheel

How to verify it
Built ptf-py3 and subsequently docker-syncd-rpc.gz with these changes. After the build sanity testing also has been done,