[Bgpcfgd] Add dependency callback logic in SRv6 Manager to handle out-of-order table processing (#21827)
- Why I did it
Previously, if the SRv6 manager received a config update that added a SRV6_MY_SID entry before its corresponding LOCATOR entry added, the SRv6 manager would not process the SRV6_MY_SID entry and print an ERR syslog. However, during testing, NVIDIA team found that it is possible to have such a scenario in normal operations, especially during config reload. So, we decide to add a dependency-based caching mechanism to handle out-of-order table processing.

This is the first use case where we will dynamically add/remove the dependency, so we also define a new unsubscribe function for the directory class. The reason why we need to remove the dependency in managers_srv6 is because the on_deps_change function will only process the cache when all dependencies are satisfied. If a locator is removed and the corresponding dependency is not removed, then the on_deps_change will never get to the satisfied condition.

Fix #21826 (Issue)