Validate docker image before uploading to registry (#21916)
Why I did it
Before pushing the newly built docker image to the registry, we add a verification step to ensure the image is valid.

Work item tracking
Microsoft ADO (number only):

How I did it
Before pushing the newly built Docker image to the registry, we add a verification step to ensure the image is valid.

How to verify it
For ref https://dev.azure.com/mssonic/build/_build/results?buildId=794495&view=results
If the validate step fails, it won't push the image to registry.