[py3] Validate docker image before uploading to registry (#21932)
Why I did it
Same as #21916, before pushing the newly built docker image to the registry, we add a verification step to ensure the image is valid.

How I did it
Same as #21916, before pushing the newly built docker image to the registry, we add a verification step to ensure the image is valid.